+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/settings/components/ErkhetSettings.tsx
@@ -25,7 +25,7 @@ const StageInErkhetConfig = lazy(() =>
 
 const PosOrderErkhetConfig = lazy(() =>
   import('~/pages/PosOrderErkhetConfigPage').then((module) => ({
-    default: module.default,
+    default: module.PosOrderErkhetConfig,
   })),
 );
