+++ frontend/plugins/mongolian_ui/src/modules/erkhet-sync/sync-erkhet-history/components/SyncErkhetHistoryDetailSheet.tsx
@@ -11,6 +11,7 @@ const stringify = (value: any) => {
     try {
       return JSON.stringify(JSON.parse(value), null, 2);
     } catch (_e) {
+      // not valid
       return value;
     }
   }
